test(msw): move MSW setup to tests directory
