+++ client/vitest.config.ts
@@ -1,10 +1,11 @@
 import { defineConfig } from "vitest/config";
 import react from "@vitejs/plugin-react";
+import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
-  plugins: [react()],
+  plugins: [react(), tsconfigPaths()],
   test: {
-    setupFiles: ["./test/msw/setup.ts"],
+    setupFiles: ["./tests/msw/setup.ts"],
     include: [
       "src/**/*.test.{ts,tsx}",
       "src/**/*.int.test.{ts,tsx}",
